updated test sepearation
