+++ package.json
@@ -17,7 +17,7 @@
     "eslint": "eslint './packages/**/*.{ts,tsx}' --fix",
     "export:storybook": "lerna run build && lerna run export:storybook",
     "prettier": "prettier --write \"packages/**/*{js,json,ts,tsx}\"",
-    "test": "NODE_ENV=test jest --verbose",
+    "test": "NODE_ENV=test jest",
     "start:components": "lerna run build:watch --stream --parallel & yarn types & lerna run start:components --stream",
     "start": "lerna run build:watch --stream --parallel & yarn types & lerna run storybook --stream",
     "types": "lerna run types",
