+++ packages/react-inbox/__mocks__/api/messages.ts
@@ -1,0 +1,27 @@
+const messages = {
+  data: {
+    messages: {
+      nodes: [
+        {
+          id: "123",
+          messageId: "1-6055153c-c0af468a3ebe8cb79665f556",
+          created: "2021-03-19T21:18:52.271Z",
+          content: {
+            title: "Template Published",
+            body: "Weekly Check-In was recently published",
+            data: {
+              clickAction:
+                "/designer/notifications/BFD3HKDEZR4478MBV5ZTGHWXXEDP",
+              tenantId: "7aff2747-d1bb-4f47-9476-3eae1646ba02",
+              triggeredBy: "Google_103278594813491371949",
+            },
+            __typename: "Content",
+          },
+          __typename: "Messages",
+        },
+      ],
+    },
+  },
+};
+
+export default messages;
